+++ sync-for-payables/CodatSyncPayables/BillPayments.cs
@@ -54,8 +54,8 @@ namespace CodatSyncPayables
         /// ### Process<br/>
         /// 1. Pass the `{billPaymentId}` to the *Delete bill payment* endpoint and store the `pushOperationKey` returned.<br/>
         /// 2. Check the status of the delete operation by checking the status of push operation either via<br/>
-        ///     1. <a href="https://docs.codat.io/introduction/webhooks/core-rules-types#push-operation-status-has-changed">Push operation webhook</a> (advised),<br/>
-        ///     2. <a href="https://docs.codat.io/sync-for-payables-api#/operations/get-push-operation">Push operation status endpoint</a>.<br/>
+        ///    1. <a href="https://docs.codat.io/introduction/webhooks/core-rules-types#push-operation-status-has-changed">Push operation webhook</a> (advised),<br/>
+        ///    2. <a href="https://docs.codat.io/sync-for-payables-api#/operations/get-push-operation">Push operation status endpoint</a>.<br/>
         /// <br/>
         ///    A `Success` status indicates that the bill payment object was deleted from the accounting platform.<br/>
         /// 3. (Optional) Check that the bill payment was deleted from the accounting platform.<br/>
@@ -66,9 +66,15 @@ namespace CodatSyncPayables
         /// ## Integration specifics<br/>
         /// Integrations that support soft delete do not permanently delete the object in the accounting platform.<br/>
         /// <br/>
-        /// | Integration | Soft Delete | Details                                                                                             |  <br/>
-        /// |-------------|-------------|-----------------------------------------------------------------------------------------------------|<br/>
-        /// | Oracle NetSuite   | No          | See <a href="/integrations/accounting/netsuite/accounting-netsuite-how-deleting-bill-payments-works">here</a> to learn more. |<br/>
+        /// | Integration | Soft Delete | Details                                                                                              |  <br/>
+        /// |-------------|-------------|------------------------------------------------------------------------------------------------------|                                                        <br/>
+        /// | Oracle NetSuite   | No          | See <a href="/integrations/accounting/netsuite/how-deleting-bill-payments-works">here</a> to learn more.  |<br/>
+        /// | QuickBooks Online | No          | -                                                                                              |<br/>
+        /// | Xero | Yes          | -                                                                                                          |<br/>
+        /// <br/>
+        /// &gt; **Supported integrations**<br/>
+        /// &gt;<br/>
+        /// &gt; This functionality is currently supported for our QuickBooks Online, Xero and Oracle NetSuite integrations.<br/>
         /// 
         /// </remarks>
         /// </summary>
@@ -130,9 +136,10 @@ namespace CodatSyncPayables
     {
         public SDKConfig Config { get; private set; }
         private const string _language = "csharp";
-        private const string _sdkVersion = "2.1.2";
-        private const string _sdkGenVersion = "2.131.1";
+        private const string _sdkVersion = "2.2.0";
+        private const string _sdkGenVersion = "2.139.1";
         private const string _openapiDocVersion = "3.0.0";
+        private const string _userAgent = "speakeasy-sdk/csharp 2.2.0 2.139.1 3.0.0 Codat.Sync.Payables";
         private string _serverUrl = "";
         private ISpeakeasyHttpClient _defaultClient;
         private ISpeakeasyHttpClient _securityClient;
@@ -157,7 +164,7 @@ namespace CodatSyncPayables
             
 
             var httpRequest = new HttpRequestMessage(HttpMethod.Post, urlString);
-            httpRequest.Headers.Add("user-agent", $"speakeasy-sdk/{_language} {_sdkVersion} {_sdkGenVersion} {_openapiDocVersion}");
+            httpRequest.Headers.Add("user-agent", _userAgent);
             
             var serializedBody = RequestBodySerializer.Serialize(request, "BillPayment", "json");
             if (serializedBody != null)
@@ -210,7 +217,7 @@ namespace CodatSyncPayables
             
 
             var httpRequest = new HttpRequestMessage(HttpMethod.Delete, urlString);
-            httpRequest.Headers.Add("user-agent", $"speakeasy-sdk/{_language} {_sdkVersion} {_sdkGenVersion} {_openapiDocVersion}");
+            httpRequest.Headers.Add("user-agent", _userAgent);
             
             
             var client = _securityClient;
@@ -258,7 +265,7 @@ namespace CodatSyncPayables
             
 
             var httpRequest = new HttpRequestMessage(HttpMethod.Get, urlString);
-            httpRequest.Headers.Add("user-agent", $"speakeasy-sdk/{_language} {_sdkVersion} {_sdkGenVersion} {_openapiDocVersion}");
+            httpRequest.Headers.Add("user-agent", _userAgent);
             
             
             var client = _securityClient;
@@ -306,7 +313,7 @@ namespace CodatSyncPayables
             
 
             var httpRequest = new HttpRequestMessage(HttpMethod.Get, urlString);
-            httpRequest.Headers.Add("user-agent", $"speakeasy-sdk/{_language} {_sdkVersion} {_sdkGenVersion} {_openapiDocVersion}");
+            httpRequest.Headers.Add("user-agent", _userAgent);
             
             
             var client = _securityClient;
@@ -354,7 +361,7 @@ namespace CodatSyncPayables
             
 
             var httpRequest = new HttpRequestMessage(HttpMethod.Get, urlString);
-            httpRequest.Headers.Add("user-agent", $"speakeasy-sdk/{_language} {_sdkVersion} {_sdkGenVersion} {_openapiDocVersion}");
+            httpRequest.Headers.Add("user-agent", _userAgent);
             
             
             var client = _securityClient;